Add DataTable Component
